チーム削除実装

+++ src/main/java/jp/co/waja/core/service/team/TeamService.java
@@ -1,16 +1,15 @@
 package jp.co.waja.core.service.team;
 
 import jp.co.waja.core.entity.Team;
-import jp.co.waja.core.model.team.TeamCreateRequest;
-import jp.co.waja.core.model.team.TeamEditRequest;
+import jp.co.waja.core.model.team.*;
+import jp.co.waja.core.repository.staff.StaffRepository;
 import jp.co.waja.core.repository.team.TeamRepository;
-import jp.co.waja.exception.NotFoundException;
+import jp.co.waja.exception.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 @Transactional
@@ -18,6 +17,9 @@ public class TeamService {
 
 	@Autowired
 	private TeamRepository teamRepository;
+
+	@Autowired
+	private StaffRepository staffRepository;
 
 	public List<Team> teams() {
 		return teamRepository.findAll();
@@ -46,8 +48,18 @@ public class TeamService {
 		return teamRepository.saveAndFlush(team);
 	}
 
-	public void delete(Long id) {
-		// Teamに１人でもStaffが所属していたら削除しない
+	public Optional<String> delete(Long id) throws NotFoundException, WrongDeleteException {
+		Team team = teamRepository.findOneById(id);
+		if (Objects.isNull(team)) {
+			throw new NotFoundException("team");
+		}
+
+		long staffCount = staffRepository.countByTeam(team);
+		if (staffCount > 0) {
+			throw new WrongDeleteException("StaffExist");
+		}
+
 		teamRepository.delete(id);
+		return Optional.ofNullable(team.getName());
 	}
 }
